Move receipt scanning to expense page
