+++ lib/views/expense_page.dart
@@ -1,6 +1,12 @@
+import 'dart:io';
+
+import 'package:cloud_functions/cloud_functions.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 import 'package:statera/models/author.dart';
 import 'package:statera/models/expense.dart';
@@ -8,6 +14,7 @@ import 'package:statera/models/item.dart';
 import 'package:statera/providers/expense_provider.dart';
 import 'package:statera/services/firestore.dart';
 import 'package:statera/utils/formatters.dart';
+import 'package:statera/utils/helpers.dart';
 import 'package:statera/viewModels/authentication_vm.dart';
 import 'package:statera/widgets/assignee_list.dart';
 import 'package:statera/widgets/author_avatar.dart';
@@ -31,6 +38,8 @@ class ExpensePage extends StatefulWidget {
 class _ExpensePageState extends State<ExpensePage> {
   AuthenticationViewModel get authVm =>
       Provider.of<AuthenticationViewModel>(context, listen: false);
+
+  final ImagePicker _picker = ImagePicker();
 
   @override
   Widget build(BuildContext context) {
@@ -83,10 +92,12 @@ class _ExpensePageState extends State<ExpensePage> {
                                     child: Text(
                                       expenseStage.name,
                                       style: TextStyle(
-                                        color: expense.isIn(expenseStage)
-                                            ? Colors.black
-                                            : Theme.of(context).textTheme.bodyText1!.color
-                                      ),
+                                          color: expense.isIn(expenseStage)
+                                              ? Colors.black
+                                              : Theme.of(context)
+                                                  .textTheme
+                                                  .bodyText1!
+                                                  .color),
                                     ),
                                   ),
                                 ),
@@ -108,12 +119,12 @@ class _ExpensePageState extends State<ExpensePage> {
                                     if (!expense.canBeUpdatedBy(
                                         authVm.user.uid)) return;
 
-                                    DateTime? newDate =
-                                        await showDatePicker(
+                                    DateTime? newDate = await showDatePicker(
                                       context: context,
                                       initialDate: DateTime.now(),
-                                      firstDate: DateTime
-                                          .fromMillisecondsSinceEpoch(0),
+                                      firstDate:
+                                          DateTime.fromMillisecondsSinceEpoch(
+                                              0),
                                       lastDate: DateTime.now().add(
                                         Duration(days: 30),
                                       ),
@@ -135,11 +146,10 @@ class _ExpensePageState extends State<ExpensePage> {
                             AuthorAvatar(
                               author: expense.author,
                               onTap: () async {
-                                if (!expense.canBeUpdatedBy(
-                                    authVm.user.uid)) return;
+                                if (!expense.canBeUpdatedBy(authVm.user.uid))
+                                  return;
 
-                                Author? newAuthor =
-                                    await showDialog<Author>(
+                                Author? newAuthor = await showDialog<Author>(
                                   context: context,
                                   builder: (context) => AuthorChangeDialog(
                                     expense: expense,
@@ -149,19 +159,17 @@ class _ExpensePageState extends State<ExpensePage> {
                                 if (newAuthor == null) return;
 
                                 expense.author = newAuthor;
-                                await Firestore.instance
-                                    .updateExpense(expense);
+                                await Firestore.instance.updateExpense(expense);
                               },
                             ),
                             Text("Assignees:"),
                             GestureDetector(
                               onTap: () {
-                                if (!expense.canBeUpdatedBy(
-                                    authVm.user.uid)) return;
+                                if (!expense.canBeUpdatedBy(authVm.user.uid))
+                                  return;
                                 showDialog(
                                   context: context,
-                                  builder: (context) =>
-                                      AssigneePickerDialog(
+                                  builder: (context) => AssigneePickerDialog(
                                     expense: expense,
                                   ),
                                 );
@@ -172,8 +180,14 @@ class _ExpensePageState extends State<ExpensePage> {
                         ),
                       ),
                       Divider(thickness: 1),
+                      if (expense.hasNoItems && !kIsWeb)
+                        ElevatedButton.icon(
+                          onPressed: () => handleScan(expense),
+                          label: Text('Upload receipt'),
+                          icon: Icon(Icons.photo_camera)
+                        ),
                       Flexible(
-                          child: expense.items.length == 0
+                          child: expense.hasNoItems
                               ? ListEmpty(text: 'Add items to this expense')
                               : ItemsList(expense: expense)),
                     ],
@@ -224,5 +238,44 @@ class _ExpensePageState extends State<ExpensePage> {
         },
       ),
     );
+  }
+
+  void handleScan(Expense expense) async {
+    final pickedFile = await _picker.pickImage(source: ImageSource.gallery);
+    if (pickedFile == null)
+      throw new Exception("Something went wrong while taking a photo");
+
+    var task = await FirebaseStorage.instance
+        .ref(pickedFile.name)
+        .putFile(File(pickedFile.path));
+
+    String url = await task.ref.getDownloadURL();
+    var getItemsFromImage =
+        FirebaseFunctions.instance.httpsCallable('getReceiptData');
+
+    var scanSuccessful = await snackbarCatch(
+      context,
+      () async {
+        var response = await getItemsFromImage({'receiptUrl': url});
+        List<dynamic> items = response.data;
+
+        items.forEach((itemData) {
+          try {
+            var item = Item(
+              name: itemData["name"] ?? "",
+              value: double.tryParse(itemData["value"].toString()) ?? 0,
+            );
+            expense.addItem(item);
+          } catch (e) {
+            print("Could not parse item $itemData: $e");
+          }
+        });
+      },
+      errorMessage: 'Something went wrong while processing your photo',
+    );
+
+    if (scanSuccessful) {
+      await Firestore.instance.updateExpense(expense);
+    }
   }
 }
